refactor: Make InjectionKeys.ts to be internal
