+++ packages/core/src/Components/Checkbox/UI/FoCheckboxGroup.vue
@@ -8,7 +8,7 @@
 
 <script setup lang="ts">
 import type { Orientation }              from '@/Shared/Types/Variants';
-import { isInCheckboxGroupInjectionKey } from '@/Components/Checkbox/Lib/InjectionKeys';
+import { isInCheckboxGroupInjectionKey } from '@/Components/Checkbox/Internal';
 import { useOrientation }                from '@/Shared/Internal/Lib';
 import { provide }                       from 'vue';
