+++ packages/core/src/Components/ListGroup/Internal/index.ts
@@ -1,0 +1,1 @@
+export * from '@/Components/ListGroup/Internal/Lib';
