+++ packages/core/src/Components/ListGroup/UI/FoListGroupItem.vue
@@ -7,7 +7,7 @@
 <script setup lang="ts">
 import type { ListGroupItemProps }          from '@/Components/ListGroup/Types/ListGroup';
 import type { Orientation }                 from '@/Shared/Types/Variants';
-import { listGroupOrientationInjectionKey } from '@/Components/ListGroup/Lib/InjectionKeys';
+import { listGroupOrientationInjectionKey } from '@/Components/ListGroup/Internal';
 import { useOrientation, useStripes }       from '@/Shared/Internal/Lib';
 import { computed, inject }                 from 'vue';
