+++ packages/core/src/Components/Menu/Internal/index.ts
@@ -1,0 +1,1 @@
+export * from '@/Components/Menu/Internal/Lib';
